Add handling for review app usage.
Setting the environment variable CLONE_URL enables this mode. It is
expected to contain a URL to a clonable git repository.

If the environment variable PLUGIN_NAME is set, a plugin review is
assumed. If the variable PROJECT_ROOT is set, a shop review is assumed.

PROJECT_ROOT should be the subfolder of the repository in which the shop
itself lies. This may be a full shopware distribution or it may only
contain the project-specific files.